+++ src/extension.ts
@@ -1,26 +1,34 @@
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
+/** @file Code Relationship Diagram extension for Visual Studio Code */
 import * as vscode from 'vscode';
 
-// This method is called when your extension is activated
-// Your extension is activated the very first time the command is executed
+/** 各国語サポート nls: National Language Support */
+import localeEn from '../package.nls.json';
+import localeJa from '../package.nls.ja.json';
+export type LocaleKeyType = keyof typeof localeEn;
+interface LocaleEntry { [key : string] : string; }
+const localeTableKey = vscode.env.language;
+const localeTable = Object.assign(localeEn, ((<{[key : string] : LocaleEntry}>{
+    ja : localeJa
+})[localeTableKey] || { }));
+const localeString = (key : string) : string => localeTable[key] || key;
+const localeMap = (key : LocaleKeyType) : string => localeString(key);
+
+/**
+ * @function 拡張機能の有効化イベント
+ * @param context extention contexest
+ */
 export function activate(context: vscode.ExtensionContext) {
+	console.log('extension "vscode-code-dependency-diagram" is now active!');
 
-	// Use the console to output diagnostic information (console.log) and errors (console.error)
-	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "vscode-code-dependency-diagram" is now active!');
-
-	// The command has been defined in the package.json file
-	// Now provide the implementation of the command with registerCommand
-	// The commandId parameter must match the command field in package.json
+	// 挨拶を表示するコマンドの登録
 	const disposable = vscode.commands.registerCommand('vscode-code-dependency-diagram.helloWorld', () => {
-		// The code you place here will be executed every time your command is executed
-		// Display a message box to the user
-		vscode.window.showInformationMessage('Hello World from vscode-code-dependency-diagram!');
+		// 右下にメッセージを表示する
+		vscode.window.showInformationMessage(localeMap('hello-message'));
 	});
-
 	context.subscriptions.push(disposable);
 }
 
 // This method is called when your extension is deactivated
-export function deactivate() {}
+export function deactivate() {
+	console.log('extension "vscode-code-dependency-diagram" is now deactivate!');
+}