+++ CHANGELOG.md
@@ -8,5 +8,6 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - 設定ファイルのファイル指定を読み込み検索する #3
 - 多言語対応を行う #2
+  - [Visual Studio Code 拡張のローカライズ対応方法](https://qiita.com/wraith13/items/8f873a1867a5cc2865a8)
 - 設定ファイルがある場合に拡張機能を起動する #1
 - Initial release